fix errror

+++ src/main/java/com/hcmutap/elearning/controller/admin/AdminCourseController.java
@@ -4,10 +4,12 @@ import com.hcmutap.elearning.dto.ClassResDTO;
 import com.hcmutap.elearning.exception.NotFoundException;
 import com.hcmutap.elearning.model.ClassModel;
 import com.hcmutap.elearning.model.CourseModel;
+import com.hcmutap.elearning.model.PointModel;
 import com.hcmutap.elearning.model.SemesterModel;
 import com.hcmutap.elearning.service.IClassService;
 import com.hcmutap.elearning.service.ICourseService;
 
+import com.hcmutap.elearning.service.IPointService;
 import com.hcmutap.elearning.service.ISemesterService;
 import jakarta.annotation.Resource;
 import org.springframework.stereotype.Controller;
@@ -27,6 +29,8 @@ public class AdminCourseController {
 	private ICourseService courseService;
 	@Resource
 	private IClassService classService;
+	@Resource
+	private IPointService pointService;
 	@Resource
 	private ISemesterService semesterService;
 
@@ -69,8 +73,13 @@ public class AdminCourseController {
 
 		try {
 			CourseModel courseModel = courseService.findById(id);
-			courseService.delete(courseModel.getFirebaseId());
-			redirectAttributes.addFlashAttribute("message", "Xóa thành công khóa học " + courseModel.getCourseId());
+			List<ClassModel> cls = classService.getClassOfCourse(id);
+			if(cls.isEmpty()) {
+				courseService.delete(courseModel.getFirebaseId());
+				redirectAttributes.addFlashAttribute("message", "Xóa thành công khóa học " + courseModel.getCourseId());
+			} else {
+				redirectAttributes.addFlashAttribute("message", "Khóa học vẫn còn lớp, không thể xóa");
+			}
 			return "redirect:/admin-management-course";
 		} catch (NotFoundException e) {
 			throw new RuntimeException(e);
@@ -119,7 +128,7 @@ public class AdminCourseController {
 		ClassModel classModel = new ClassModel();
 		classModel.setCourseId(classRes.getCourseId());
 		classModel.setClassName(classRes.getClassName());
-		classModel.setClassId(classModel.getCourseId()+"-"+classModel.getClassName());
+		classModel.setClassId(classModel.getCourseId()+"-"+classModel.getClassName()+"-"+classRes.getSemesterId());
 		classModel.setDayOfWeek(classRes.getDayOfWeek());
 		classModel.setTimeStart(classRes.getTimeStart());
 		classModel.setTimeEnd(transferTime(classRes.getTimeStart(), classRes.getTimeStudy()));
@@ -214,11 +223,15 @@ public class AdminCourseController {
 
 	@GetMapping("/admin-management/deleteClass")
 	public String deleteClass (@RequestParam("id") String id, final RedirectAttributes redirectAttributes) {
-
 		try {
 			ClassModel classModel = classService.findById(id);
-			classService.delete(classModel.getFirebaseId());
-			redirectAttributes.addFlashAttribute("message", "Xóa thành công lớp " + classModel.getClassName());
+			try {
+				List<PointModel> points = pointService.getListStudentByClassId(id);
+				redirectAttributes.addFlashAttribute("message", "Vẫn còn sinh viên trong lớp, không thể xóa!");
+			} catch (Exception e) {
+				classService.delete(classModel.getFirebaseId());
+				redirectAttributes.addFlashAttribute("message", "Xóa thành công lớp " + classModel.getClassName());
+			}
 			return "redirect:/admin-management/update-course?id=" + classModel.getCourseId();
 		} catch (NotFoundException e) {
 			throw new RuntimeException(e);
